backend/requirements.txt
# Core FastAPI and async
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MongoDB and Beanie ODM
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development tools
python-dotenv==1.0.0

# For your scraper integration
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
lxml==4.9.3

# Logging and monitoring
python-json-logger==2.0.7